Play Jeopardy!
This is a project from Codecademy where I used a database of Jeopardy! gameshow data to allow users to select an air date to play using the questions from a particular episode, select their own squares, and then try to get the right answer (in the form of a question, of course).